producer index to differentiate producers
at the moment producer id is generated only once and used for every producer for specific Kafka cluster. this changes will create producer with index so that we can differentiate them on monitoring or other tooling we could use. this is the current behaviour of Kafka client to add integer suffix for producer id