Fix NPE in `AllowListService` due to missing LoLA limit config (#291)
We have two configurations: static via the env. var and dynamic via the
feature toggle to control this behavior.  If the first one is not set, the
LoLA connections cannot be open, because we did not create the node cache.

Fixed by checking the static setting before trying to use the node cache.